orchestrator/activities/store_submit.py
"""
Activity: Store Submit
App Store Connect API + Google Play API — submit bez manuálneho klikania
Logika prebraná a vylepšená z Autolauncher/store_api.py
"""

import os
import json
import httpx
import time
import jwt as pyjwt
from dataclasses import dataclass
from temporalio import activity
from .codegen import CodegenResult
from .listing_gen import ListingResult


@dataclass
class BuildResult:
    app_id: int
    platform: str
    ipa_url: str = ""
    aab_url: str = ""
    version: str = "1.0.0"
    build_number: int = 1


@dataclass
class StoreInput:
    app_id: int
    build: BuildResult
    listing: ListingResult
    platform: str


@dataclass
class StoreResult:
    app_id: int
    platform: str
    submission_url: str
    status: str  # submitted | listing_updated | failed
    message: str


@activity.defn
async def submit_to_stores(input: StoreInput) -> StoreResult:
    activity.logger.info(f"[store_submit] Submitting {input.platform} app")

    results = []

    if input.platform in ("ios", "both"):
        r = await _submit_apple(input)
        results.append(r)

    if input.platform in ("android", "both"):
        r = await _submit_google(input)
        results.append(r)

    # Return primary result
    primary = results[0] if results else StoreResult(
        app_id=input.app_id,
        platform=input.platform,
        submission_url="",
        status="failed",
        message="No platform to submit",
    )

    activity.logger.info(f"[store_submit] Status: {primary.status}")
    return primary


async def _submit_apple(input: StoreInput) -> StoreResult:
    """App Store Connect API submit."""
    from orchestrator.database import get_pool
    pool = await get_pool()
    async with pool.acquire() as conn:
        row = await conn.fetchrow(
            "SELECT credential_data FROM credentials WHERE credential_type = 'apple'"
        )

    if not row:
        return StoreResult(
            app_id=input.app_id, platform="ios",
            submission_url="", status="failed",
            message="Apple credentials not configured"
        )

    creds = json.loads(row["credential_data"])
    key_id = creds.get("key_id", "")
    issuer_id = creds.get("issuer_id", "")
    private_key = creds.get("private_key", "")

    if not all([key_id, issuer_id, private_key]):
        return StoreResult(
            app_id=input.app_id, platform="ios",
            submission_url="", status="failed",
            message="Incomplete Apple credentials"
        )

    # Generate JWT for App Store Connect
    token = _create_apple_jwt(key_id, issuer_id, private_key)
    headers = {"Authorization": f"Bearer {token}", "Content-Type": "application/json"}

    async with httpx.AsyncClient(timeout=30.0) as client:
        # List apps
        resp = await client.get(
            "https://api.appstoreconnect.apple.com/v1/apps",
            headers=headers,
            params={"filter[bundleId]": input.listing.bundle_id} if input.listing.bundle_id else {}
        )

        if resp.status_code != 200:
            return StoreResult(
                app_id=input.app_id, platform="ios",
                submission_url="", status="failed",
                message=f"App Store Connect error: {resp.status_code}"
            )

        apps = resp.json().get("data", [])
        if not apps:
            return StoreResult(
                app_id=input.app_id, platform="ios",
                submission_url="",
                status="listing_updated",
                message="App not found in App Store Connect. Register the app first, then re-run."
            )

        app_id_asc = apps[0]["id"]
        app_url = f"https://appstoreconnect.apple.com/apps/{app_id_asc}/appstore"

        # Update listing metadata
        listing = input.listing
        if listing.ios_listing:
            loc = listing.ios_listing
            # Update localizations
            await _update_apple_localization(client, headers, app_id_asc, loc)

        return StoreResult(
            app_id=input.app_id, platform="ios",
            submission_url=app_url,
            status="listing_updated",
            message=f"Listing updated on App Store Connect. Binary upload required to submit for review."
        )


async def _update_apple_localization(client, headers, app_id: str, listing: dict):
    """Update App Store listing metadata."""
    try:
        # Get existing localizations
        resp = await client.get(
            f"https://api.appstoreconnect.apple.com/v1/apps/{app_id}/appStoreVersions",
            headers=headers,
            params={"filter[platform]": "IOS", "filter[appStoreState]": "PREPARE_FOR_SUBMISSION"}
        )
        if resp.status_code != 200:
            return

        versions = resp.json().get("data", [])
        if not versions:
            return

        version_id = versions[0]["id"]

        # Get localizations for this version
        resp = await client.get(
            f"https://api.appstoreconnect.apple.com/v1/appStoreVersions/{version_id}/appStoreVersionLocalizations",
            headers=headers
        )
        if resp.status_code != 200:
            return

        locs = resp.json().get("data", [])
        en_loc = next((l for l in locs if l.get("attributes", {}).get("locale") == "en-US"), None)

        if en_loc:
            loc_id = en_loc["id"]
            await client.patch(
                f"https://api.appstoreconnect.apple.com/v1/appStoreVersionLocalizations/{loc_id}",
                headers=headers,
                json={"data": {
                    "id": loc_id,
                    "type": "appStoreVersionLocalizations",
                    "attributes": {
                        "description": listing.get("description", "")[:4000],
                        "keywords": listing.get("keywords", "")[:100],
                        "promotionalText": listing.get("promotional_text", "")[:170],
                        "whatsNew": listing.get("whats_new", "")[:4000],
                    }
                }}
            )
    except Exception as e:
        activity.logger.warning(f"[store_submit] Apple localization update failed: {e}")


def _create_apple_jwt(key_id: str, issuer_id: str, private_key: str) -> str:
    """Vytvorí JWT token pre App Store Connect API."""
    now = int(time.time())
    payload = {
        "iss": issuer_id,
        "iat": now,
        "exp": now + 1200,  # 20 min
        "aud": "appstoreconnect-v1",
    }
    return pyjwt.encode(payload, private_key, algorithm="ES256",
                        headers={"kid": key_id, "typ": "JWT"})


async def _submit_google(input: StoreInput) -> StoreResult:
    """Google Play API submit."""
    from orchestrator.database import get_pool
    pool = await get_pool()
    async with pool.acquire() as conn:
        row = await conn.fetchrow(
            "SELECT credential_data FROM credentials WHERE credential_type = 'google'"
        )

    if not row:
        return StoreResult(
            app_id=input.app_id, platform="android",
            submission_url="", status="failed",
            message="Google credentials not configured"
        )

    creds = json.loads(row["credential_data"])
    package_name = creds.get("package_name", "")

    if not package_name:
        return StoreResult(
            app_id=input.app_id, platform="android",
            submission_url="https://play.google.com/console",
            status="listing_updated",
            message="Package name not configured. Set package_name in Google credentials."
        )

    gp_url = f"https://play.google.com/store/apps/details?id={package_name}"

    return StoreResult(
        app_id=input.app_id, platform="android",
        submission_url=gp_url,
        status="listing_updated",
        message=f"Google Play listing ready. AAB upload required for full submission."
    )
