requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.30.0
temporalio==1.7.1
anthropic==0.49.0
openai==1.58.0
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
sentry-sdk==2.19.2
prometheus-client==0.21.0
